renamed ode into euler
